+++ platformio.ini
@@ -413,11 +413,11 @@ extra_scripts             = pre:pre_extra_script.py
 ; This may generate builds which cannot be run, so do not upload to a node.
 [env:memanalyze_ESP8266]
 platform                  = ${normal_252.platform}
-board                     = ${esp8266_4M.board}
-board_build.f_cpu         = ${esp8266_4M.board_build.f_cpu}
-board_build.flash_mode    = ${esp8266_4M.board_build.flash_mode}
-build_unflags             = ${esp8266_4M.build_unflags}
-build_flags               = ${esp8266_4M.build_flags} ${normal_252.build_flags} -DPLUGIN_BUILD_CUSTOM -w
+board                     = ${esp8266_4M1M.board}
+board_build.f_cpu         = ${esp8266_4M1M.board_build.f_cpu}
+board_build.flash_mode    = ${esp8266_4M1M.board_build.flash_mode}
+build_unflags             = ${esp8266_4M1M.build_unflags}
+build_flags               = ${esp8266_4M1M.build_flags} ${normal_252.build_flags} -DPLUGIN_BUILD_CUSTOM -w
 
 extra_scripts             = pre:pre_memanalyze.py
