+++ src/ESPEasy_common.cpp
@@ -1,0 +1,13 @@
+#include "ESPEasy_common.h"
+
+/*********************************************************************************************\
+   Bitwise operators
+  \*********************************************************************************************/
+bool getBitFromUL(uint32_t number, byte bitnr) {
+  return (number >> bitnr) & 1UL;
+}
+
+void setBitToUL(uint32_t& number, byte bitnr, bool value) {
+  uint32_t newbit = value ? 1UL : 0UL;
+  number ^= (-newbit ^ number) & (1UL << bitnr);
+}
