+++ src/ESPEasy_fdwdecl.h
@@ -23,7 +23,7 @@
 
 // Forward declaration
 struct ControllerSettingsStruct;
-static String getUnknownString();
+String getUnknownString();
 void scheduleNextDelayQueue(unsigned long id, unsigned long nextTime);
 String LoadControllerSettings(int ControllerIndex, ControllerSettingsStruct& controller_settings);
 String get_formatted_Controller_number(int controller_index);
